feat(model): add LSTM model and callbacks for training pipeline
- Introduced a new LSTM model implementation in lstm.py
- Added functions to create LSTM model and associated callbacks
- Updated train.py to utilize the new LSTM model and callbacks
